requirements.txt

Django>=3.2,<4.2.7
psycopg2-binary>=2.8
gunicorn>=21.2.0
django-environ>=0.11.2
python-dotenv>=1.0.1
djangorestframework>=3.15.2
whitenoise>=6.6.0